add manifest for persistent volumes for database
This manifest defines three Kubernetes persistent volumes, each
using a distinct Azure File Share; these will be bound via
persistent volume claims to the three replicas in the stateful set
of pods that will run MongoDB.

Using persistent volumes ensures that the underlying filesystems
in the File Shares are maintained across redeployments of the
MongoDB database; otherwise, data would be lost and the database
would likely enter a corrupt state.

The Azure File Shares are expected to exist, and to be named with
a common base name followed by an ordinal number.  The base name
will be interpolated with a GitHub Actions secret which will replace
the FILE_SHARE_BASE_NAME string.

The volume's reclaim policy is set to the "retain" mode to ensure
they are not deleted when the claims to them are released, should
the MongoDB pods be stopped, although this should not happen since
these volumes are deployed statically not dynamically.