add dispatch workflow to deploy persistent volumes
As the Kubernetes persistent volumes required for the MongoDB database
should be deployed in advance of database and application deployments,
and persist across them, add a dispatch workflow to perform the
deployment of the volumes only when triggered by an administrative
dispatch event.

This GitHub Actions workflow expects an Azure Kubernetes Service cluster
to exist and for its name and resource group to be configured as
GitHub Actions secrets.

The workflow also expects Azure File Shares for the volumes to exist
with a common base name which is also configured in a GitHub Actions
secret, as well as the Azure login credentials.

After deployment of the persistent volumes manifest, the workflow
uses kubectl to wait for the volumes' statuses to reach the
available phase.