add manifest for statefulset for database
This manifest defines a Kubernetes stateful set of three replica pods,
each running MongoDB and bound to a persistent volume which is mounted
at /data/db so the database files will be persisted across pod
redeployments; otherwise, data would be lost and the database would
likely enter a corrupt state.

The use of a stateful set ensures that when a pod fails, the replacement
pod will be bound to the same volume, and will have the same identifier
so application clients and other pods' database processes can connect
to it at the same address.  The default update strategy for a stateful
set is the RollingUpdate strategy, so each pod is updated in turn
without overlap.

The manifest also defines a headless service by setting the cluster IP
value to "None"; this form of service is required by stateful sets
so that the pods are assigned fixed network identities.

The MongoDB database processes bind to all available addresses and
listen on the standard MongoDB port 27017, so they will be available
at, for instance, mongo-0.mongo:27017, mongo-1.mongo:27017, etc.