add dispatch workflow to deploy MongoDB database
As the Kubernetes stateful set of pods required for the MongoDB database
should be deployed in advance of application deployments, and persist
across them, add a dispatch workflow to perform the deployment of the
stateful set only when triggered by an administrative dispatch event.

This GitHub Actions workflow expects an Azure Kubernetes Service cluster
to exist and for its name and resource group to be configured as
GitHub Actions secrets.  It also expects an Azure Docker container
registry to exist and for its name and credentials to be configured
as GitHub Actions secrets.

As well, the workflow expects a set of three persistent volumes to exist
to be claimed by the stateful set's pods.

After deployment of the stateful set manifest, the workflow uses kubectl
to wait for the pods to reach the ready condition.

Once that state has been reached, the workflow job creates or updates
the MongoDB replica set (which is a distinct concept from a Kubernetes
replica set), using kubectl to execute the mongosh interactive shell.
If the replica set has been initiated previously MongoDB will report
an error message but proceed without halting.

All three pods defined by the manifest will be registered in the
MongoDB replica set, using their internal DNS addresses of mongo-1.mongo,
mongo-2.mongo, etc.  If any have been registered previously, MongoDB
will report the issue and ignore the new registrations.

After waiting for the registrations to complete, the job uses kubectl
and mongosh to check the status of the MongoDB replica set; there should
be one primary database replica and two secondaries.

Finally, the job creates the default "db" database in MongoDB, if it
does not exist, by inserting a timestamp row.  In the case where
the database has been deployed before, this step updates the timestamp
instead.