add manifest for pod deployment for app
This manifest defines a Kubernetes deployment for three application pods
with the "role" label of "frontend", to which the load balanacer
service will direct traffic on port 8081.

The rolling update strategy is used so pods will be redeployed incrementally
with no downtime.  No more than one extra pod is allowed at any time,
and no more than one pod may be unavailable at any time.

Liveness and readiness checks are defined using the application's
health status URL path.

The Docker image for the application is expected to be available in
an Azure container registry whose name should be interpolated into
the manifest in place of the REGISTRY_LOGIN_SERVER string by the
deployment GitHub Actions workflow.

The workflow should also interpolate a series of configuration values
for the application itself, including the secrets used to sign the
application's cookies and the login and password for the application.