add default branch push workflow to deploy app
This GitHub actions workflow runs on every push merged into the "demo"
branch, and will build, test, and deploy the application, with an
explicit manual approval required after an initial canary deployment
to 20% of the application's pods.

This GitHub Actions workflow expects an Azure Kubernetes Service cluster
to exist and for its name and resource group to be configured as
GitHub Actions secrets.  It also expects an Azure Docker container
registry to exist and for its name and credentials to be configured
as GitHub Actions secrets.

The general Azure credentials must also be configured as GitHub Actions
secrets, as well as a set of application configuration values, including
the application's cookie-signed secrets and the login and password it
should require from users.

The workflow's initial test job is derived from the existing upstream
mongo-express project's standard CI test workflow.

The subsequent job builds a Docker image of the application and
pushes it into the Azure container registry.

If this succeeds, the next job deploys the built Docker image to a
canary pod, to which the load balancer will route 20% of the overall
application traffic.  Note that the azure/k8s-deploy action's canary
strategy calculates the number of canary and baseline pods to be added
to the existing set such that the requested input traffic percentage
does not match the actual percentage.  Because the manifest defines
a deployment of three pods, given an input percentage of 34%, the
action calculates floor(34 * 3 / 100) = 1, so one baseline and one
canary replica are then added to the existing deployment.  This results
in an actual percentage of traffic routed to the canary pod of 20%.
Note also that the baseline pod simply runs the existing deployed
image of the application, not the new image.

After the canary deployment of the manifest and the new Docker image,
the workflow uses kubectl to wait for the new canary and baseline pods
to reach the ready condition.

Upon a successful canary deployment, the workflow then pauses for
explicit manual review of the health of the application and approval
to proceed with a full deployment, assuming the "demo" environment
has been configured with required reviewers for any deployment.

If approval is not granted, the next job is skipped and the final
job runs to roll back the canary deployment by deleting the canary
and baseline pods.

If approval is granted, however, the final job is skipped and the
workflow instead proceeds with a job to complete the full deployment.
The normal pods will be replaced with ones running the new Docker
image, according to the rolling update strategy defined in the
manifest, and the baseline and canary pods will be deleted.

Both of these jobs use kubectl to wait for the canary and baseline
pods to be deleted, and for the remaining pods to reach the ready
condition.  In the rollback case, these pods should be untouched
and already in this condition.